requirement.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.13.1
torch==2.2.2
transformers==4.41.2
tqdm==4.66.4
